Using object instead of list for sensors
with this commit I switched to using an object for all the sensors. This
way sensors can be more easily read out as you can address them by their
id, rather then iterating over the list of sensors and finding the one
you want to read out.
Yeah, technically you could reference the sensor by its position in the
list, but as that might change I feel this way is not really robust.

Also fixed a typo in the humidity sensors.